Fixed naming errors in zone_dialog
